+++ src/dapp/constants.ts
@@ -16,6 +16,8 @@ export const WALLET_PROVIDER_CONNECT_URL = "/hook/login";
 export const WALLET_PROVIDER_DISCONNECT_URL = "/hook/logout";
 export const WALLET_PROVIDER_SEND_TRANSACTION_URL = "/hook/transaction";
 export const WALLET_PROVIDER_SIGN_TRANSACTION_URL = "/hook/sign";
+export const WALLET_PROVIDER_CALLBACK_PARAM = "walletProviderStatus";
+export const WALLET_PROVIDER_CALLBACK_PARAM_TX_SIGNED = "transactionsSigned";
 
 // This constant represents the minimum version in which the Elrond Ledger App doesn't support anymore regular
 // transactions' signing, and uses transaction's hash signing instead