+++ CHANGELOG.md
@@ -33,6 +33,16 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - [Added Extension Provider for dapp integration #59](https://github.com/ElrondNetwork/elrond-sdk-erdjs/pull/59)
 
+- [Refactored WalletProvider #60](https://github.com/ElrondNetwork/elrond-sdk-erdjs/pull/60)
+   - Refactored wallet provider to be compatible with all browsers
+   - Added `signTransactions` so users can sign transaction batches
+   - Added `getTransactionsFromWalletUrl` endpoint so users can retrieve the decoded `Transaction` object from the `callbackUrl`
+  
+   - ### Breaking changes:
+     - The WalletProvider no longer communicates with the ElrondWallet through an iframe
+     - The above means that we no longer have a way to get some info - so `getAddress` will throw `ErrNotImplemented`
+     - Removed `customId` from constructor
+
 ## [6.6.2]
 -   [Quickfix - added custom id to iframe creation #55](https://github.com/ElrondNetwork/elrond-sdk-erdjs/pull/55)